update examples remove duplicate code
