+++ examples/event/main.go
@@ -17,7 +17,7 @@ import (
 	// "go-vgo/robotgo"
 )
 
-func main() {
+func event() {
 	////////////////////////////////////////////////////////////////////////////////
 	// Global event listener
 	////////////////////////////////////////////////////////////////////////////////
@@ -54,4 +54,8 @@ func main() {
 
 	// stop AddEvent
 	// robotgo.StopEvent()
+}
+
+func main() {
+	event()
 }
