+++ examples/key/main.go
@@ -17,7 +17,7 @@ import (
 	// "go-vgo/robotgo"
 )
 
-func main() {
+func key() {
 	////////////////////////////////////////////////////////////////////////////////
 	// Control the keyboard
 	////////////////////////////////////////////////////////////////////////////////
@@ -71,4 +71,8 @@ func main() {
 	if err == nil {
 		fmt.Println(text)
 	}
+}
+
+func main() {
+	key()
 }
