+++ examples/mouse/main.go
@@ -17,7 +17,7 @@ import (
 	// "go-vgo/robotgo"
 )
 
-func main() {
+func mouse() {
 	////////////////////////////////////////////////////////////////////////////////
 	// Control the mouse
 	////////////////////////////////////////////////////////////////////////////////
@@ -58,4 +58,8 @@ func main() {
 		robotgo.MoveMouse(800, i)
 	}
 
+}
+
+func main() {
+	mouse()
 }
