+++ examples/window/main.go
@@ -17,7 +17,7 @@ import (
 	// "go-vgo/robotgo"
 )
 
-func main() {
+func window() {
 	////////////////////////////////////////////////////////////////////////////////
 	// Window Handle
 	////////////////////////////////////////////////////////////////////////////////
@@ -98,4 +98,8 @@ func main() {
 
 	// close current Window
 	robotgo.CloseWindow()
+}
+
+func main() {
+	window()
 }
